Task #3410: Committing changes related to merging the trunk into my task branch

+++ branches/MAC-Swleveladapt-Task3410/CEP/GSM/bremen/src/pipeline.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python
 import monetdb.sql as db
-from src.errors import SourceException
+from src.errors import SourceException, ImageStateError
 from src.gsmconnectionmanager import GSMConnectionManager
 from src.gsmlogger import get_gsm_logger
-from src.sqllist import get_sql
+from src.sqllist import get_sql, get_svn_version
 from src.grouper import Grouper
 from src.updater import run_update
 import logging
@@ -71,16 +71,24 @@ class GSMPipeline(object):
             grouper.cleanup()
         self.conn.execute(get_sql("GroupFill"))
 
-    def process_image(self, image_id):
+    def process_image(self, image_id, sources_loaded=False):
         """
         Process single image.
+        @sources_loaded: True if there are records in the extractedsources
+        already.
         """
         self.conn.start()
-        band = self.conn.exec_return("select band from images"\
-                                     " where imageid = %s;" % image_id)
+        status, band, stokes = self.conn.exec_return("""
+        select status, band, stokes
+          from images
+         where imageid = %s;""" % image_id, single_column=False)
+        if status == 1:
+            raise ImageStateError('Image %s in state 1 (Ok). Cannot process' %
+                                  image_id)
         self.conn.execute("delete from temp_associations;")
-        self.conn.execute(get_sql('insert_extractedsources', image_id))
-        self.conn.execute(get_sql('insert dummysources', image_id))
+        if not sources_loaded:
+            self.conn.execute(get_sql('insert_extractedsources', image_id))
+            self.conn.execute(get_sql('insert dummysources', image_id))
         self.conn.execute(get_sql('Associate point',
                                       image_id, math.sin(0.025), 1.0))
         self.conn.execute_set(get_sql('Associate extended',
@@ -108,4 +116,11 @@ class GSMPipeline(object):
         #inserting new sources
         self.conn.execute_set(get_sql('Insert new sources', image_id))
         self.conn.execute_set(get_sql('Join extended', image_id))
+        #update image status and save current svn verion.
+        self.conn.execute("""
+update images
+   set status = 1,
+       process_date = current_timestamp,
+       svn_version = %s
+ where imageid = %s""" % (get_svn_version(), image_id))
         self.conn.commit()
